requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
amqp==5.3.1
asgiref==3.8.1
asttokens==3.0.0
attrs==24.3.0
backcall==0.2.0
beautifulsoup4==4.12.3
billiard==4.2.1
bleach==6.2.0
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
dateparser==1.2.0
decorator==5.1.1
defusedxml==0.7.1
Django==5.1.1
django-celery-results==2.5.1
django-cors-headers==4.6.0
django-crispy-forms==2.3
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
docopt==0.6.2
et-xmlfile==1.1.0
executing==2.1.0
fastjsonschema==2.21.1
filelock==3.15.4
frozenlist==1.5.0
fsspec==2024.9.0
gunicorn==23.0.0
idna==3.8
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kombu==5.4.2
llvmlite==0.43.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
more-itertools==10.5.0
mpmath==1.3.0
multidict==6.1.0
mysqlclient==2.2.6
nbclient==0.10.2
nbconvert==7.16.4
nbformat==5.10.4
networkx==3.3
numba==0.60.0
numpy==2.0.2
openai-whisper==20231117
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pillow==11.0.0
pipreqs==0.5.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.1
pure_eval==0.2.3
pycryptodome==3.21.0
Pygments==2.18.0
python-binance==1.0.26
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
pywin32==308
pyzmq==26.2.0
redis==5.2.1
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rpds-py==0.22.3
setuptools==74.1.2
six==1.17.0
soupsieve==2.6
sqlparse==0.5.1
stack-data==0.6.3
sympy==1.13.2
tiktoken==0.7.0
tinycss2==1.4.0
torch==2.4.1
tornado==6.4.2
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.1
yarg==0.1.9
yarl==1.18.3
